+++ src/main/java/com/snowhite/server/web/dto/web/response/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.snowhite.server.web.dto;
+package com.snowhite.server.web.dto.web.response;
 
 import lombok.Builder;
