+++ src/main/java/com/snowhite/server/websocket/dto/request/GameJoinRequest.java
@@ -1,4 +1,4 @@
-package com.snowhite.server.web.dto.websocket.request;
+package com.snowhite.server.websocket.dto.request;
 
 public record GameJoinRequest(
         Long playerId
