+++ src/main/java/com/snowhite/server/websocket/dto/response/SimpleMessageResponse.java
@@ -1,4 +1,4 @@
-package com.snowhite.server.web.dto.websocket.response;
+package com.snowhite.server.websocket.dto.response;
 
 public record SimpleMessageResponse(
         String message
