+++ src/main/java/com/snowhite/server/websocket/handler/GameWebSocketHandler.java
@@ -1,4 +1,4 @@
-package com.snowhite.server.web.websocket;
+package com.snowhite.server.websocket.handler;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.snowhite.server.domain.session.Game;
 import com.snowhite.server.security.jwt.JwtProvider;
 import com.snowhite.server.service.GameService;
-import com.snowhite.server.web.dto.websocket.response.SimpleMessageResponse;
+import com.snowhite.server.websocket.dto.response.SimpleMessageResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.ReactiveRedisTemplate;
 import org.springframework.stereotype.Component;
